PR review. Part one.
